chore : camera - hooks edit
